Use a custom sorting approach with Deft
Deft presumably is intended to mimic Notational Velocity. One very nice feature of Notational Velocity was that it
sorted your search in a human-friendly fashion. That is, if the title of a note matched what you searched for, it would
come up first before all other kind of matches, which would be sorted by modified time. Deft, on the other hand, has an
awkward "sort all titles alphabetically" vs "sort all files by modified time" approach. I wanted the middle ground, so I
implemented a custom sort for the view. Note that Deft presorts all files in order to speed up the filter-and-display
process, which is commendable; my sort happens after the search.