Fixed Alchemist some interaction with Docker with regard to definition discovery
Note that this means Alchemist will now (correctly) attempt to start its server within the Docker container or SSH
container or wherever. The server requires the same path as on your local computer. With a Docker container, this means
sharing the $HOME/.emacs.d/ directory (or the alchemist directory underneath that).